extract extended info
