+++ src/components/layout/InfoForToday.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { SeeMoreArrow } from 'components/layout/SeeMoreArrow';
-import { CalendarDataType, getDayOfWeek } from 'utils/date';
-import { makeStartCase, formatCalendarNotes } from 'utils/text';
-import { copy } from 'consts/copy';
+import { ExtendedInfo } from 'components/layout/ExtendedInfo';
+import { CalendarDataType } from 'utils/date';
+import { makeStartCase } from 'utils/text';
 
 export default function InfoForToday({ data }: { data: CalendarDataType }) {
   function handleClick() {
@@ -16,8 +16,6 @@ export default function InfoForToday({ data }: { data: CalendarDataType }) {
   }
 
   const [showMoreInfo, setShowMoreInfo] = useState(false);
-  const dayOfWeek = getDayOfWeek();
-  const [intro, linkToReadings] = copy.sundayReadings.split('[html]');
 
   return (
     <div className="font-light md:mb-4 mt-2 text-sm more-info">
@@ -36,38 +34,7 @@ export default function InfoForToday({ data }: { data: CalendarDataType }) {
         </div>
         <SeeMoreArrow rotated={showMoreInfo} />
       </div>
-      {showMoreInfo && (
-        <div>
-          <ul className="mt-2">
-            {data?.passages?.map((passage, index) => (
-              <li key={index}>
-                <a
-                  href={`https://wbiblii.pl/szukaj/${passage}`}
-                  className="hover:underline"
-                >
-                  {passage}
-                </a>
-              </li>
-            ))}
-          </ul>
-          {dayOfWeek === 'niedziela' && (
-            <div>
-              {intro}{' '}
-              <a
-                href="https://starokatolicy.eu/czytania-liturgiczne/"
-                className="font-medium"
-              >
-                {linkToReadings}
-              </a>
-            </div>
-          )}
-          <ul className="mt-1">
-            {data?.notes?.map((note, index) => (
-              <li key={index}>{formatCalendarNotes(note)}</li>
-            ))}
-          </ul>
-        </div>
-      )}
+      {showMoreInfo && <ExtendedInfo data={data} />}
     </div>
   );
 }
